requirements.txt
fastapi
uvicorn[standard]
notion-client
python-dotenv
requests
